project_c_llm_orchestration/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
langchain>=1.0.0
langchain-google-genai>=4.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
